feat: show quick win label in sheet overview
